% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_chart.R
\name{generic_chart_ui}
\alias{generic_chart_ui}
\title{Generic Chart Shiny Module (UI)}
\usage{
generic_chart_ui(id, title = "", sliderWidth = 25)
}
\arguments{
\item{id}{Character string. Module namespace identifier.}

\item{title}{Character string, title displayed at the top of the chart box.}

\item{sliderWidth}{A numeric value between 25 and 100. Sidebar opening width in percentage. 25\%, means the card sidebar will take 25\% of the card width, when opened. 
Default is \code{25}.}
}
\description{
User interface for the generic chart Shiny module.

The UI provides:
\itemize{
  \item A configurable sidebar for plot options
  \item A main panel displaying interactive charts and summary tables
  \item Download controls for exporting charts (PNG / HTML)
}

This UI is designed to work with \code{generic_chart_server} and relies on
\code{bs4Dash} plotly::layout components.
}
