% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome.R
\name{welcome_server}
\alias{welcome_server}
\title{POC shiny module server}
\usage{
welcome_server(id, lang = NULL)
}
\arguments{
\item{id}{id}

\item{lang}{lang a reactive version of the language. Default is \code{NULL} (optional)}
}
\description{
POC shiny module server
}
