% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdishinyr.R
\docType{package}
\name{fdishinyr}
\alias{fdishinyr-package}
\alias{fdishinyr}
\title{FDI shiny utils}
\description{
FDI shiny utils
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdiwg/fdishinyr}
  \item Report bugs at \url{https://github.com/fdiwg/fdishinyr/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
